docker/update-table-reference/pyproject.toml
[tool.poetry]
name = "updatereference"
version = "0.1.0"
description = ""
authors = ["Quang Truong <truong.quang@scheidt-bachmann-st.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
GitPython = "^3.1.43"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"